feat(firmware2): port oscillator hard sync + retrigger phase (gap queue #1+#2)
Oscillator hard sync (C voice.cpp:1100-1106/1171-1240/2400-2430, render_wave.h:25-90):
- Sound: oscillatorSync flag + renderingOscillatorSyncCurrently() (sound.h:143, sound.cpp:2112-2121).
- Voice.render: doingOscSync capture of osc A's per-part phases before rendering; subtractive
  source-0 per-part final-increment collection (including the inactive-source-0 collect+advance
  path, voice.cpp:1181-1199/1993-1996); osc B renders with per-part sync resetter args and skips
  parts whose osc-A pitch was too high (voice.cpp:2403-2406); ringmod osc B syncs to osc A
  (voice.cpp:1317-1340).
- Oscillator: renderOscSync ported (render_wave.h) — raw wave segments between hard-sync resets
  with the half-sine crossover-sample blend, then the amplitude pass
  (applyAmplitudeVectorToBuffer, oscillator.cpp:510-529) — wired into the SINE/TRIANGLE-table/
  SAW-table/SQUARE-table paths (callRenderWave sync branch, oscillator.cpp:476-498) and the crude
  triangle loop (oscillator.cpp:177-196). Crude saw already had it.
- Subtractive osc render now also passes the PULSE WIDTH param (voice.cpp:2423 — was hardcoded 0).

Retrigger phase (C sound.h:156-157, vups:79-82, voice.cpp:319-327, retrigger_phase.h:49-58):
- Sound: oscRetriggerPhase[2]/modulatorRetriggerPhase[2], raw uint32, 0xFFFFFFFF=off (C default;
  model defaults fixed 0→-1 to match).
- Voice.noteOn: random phases (C randomizeOscPhases voice.cpp:399-411 — modulator phases now
  random for FM like the C, pinned to 0 under the test override) then retrigger pinning.
- Bridge: trigger-path propagation of noteOn-applied state (renderVoicesFw2 propagates too late
  for noteOn); factory maps model oscillatorSync; parser retrigPhase reads hardened to
  Long.parseLong (raw uint32 > 2^31 overflowed Integer.parseInt).

The UI "Oscillator Sync" toggle and "Retrig Phase" controls are now real.
Tests: OscSyncRetrigPhaseTest (determinism via retrig; hard-sync periodicity lock 0.55→0.99).
Suites: default 339/0, slow 416/0.

+++ src/main/java/org/chuck/deluge/xml/DelugeXmlParser.java
@@ -1134,7 +1134,7 @@ public class DelugeXmlParser {
       NodeList rpNodes = osc1.getElementsByTagName("retrigPhase");
       if (rpNodes.getLength() > 0) {
         try {
-          synth.setOsc1RetrigPhase(Integer.parseInt(rpNodes.item(0).getTextContent().trim()));
+          synth.setOsc1RetrigPhase((int) Long.parseLong(rpNodes.item(0).getTextContent().trim()));
         } catch (NumberFormatException e) {
           LOG.log(Level.FINE, "NumberFormatException parsing XML attribute", e);
         }
@@ -1182,7 +1182,7 @@ public class DelugeXmlParser {
       NodeList rpNodes2 = osc2.getElementsByTagName("retrigPhase");
       if (rpNodes2.getLength() > 0) {
         try {
-          synth.setOsc2RetrigPhase(Integer.parseInt(rpNodes2.item(0).getTextContent().trim()));
+          synth.setOsc2RetrigPhase((int) Long.parseLong(rpNodes2.item(0).getTextContent().trim()));
         } catch (NumberFormatException e) {
           LOG.log(Level.FINE, "NumberFormatException parsing XML attribute", e);
         }
@@ -1810,7 +1810,7 @@ public class DelugeXmlParser {
       NodeList rpNodes = mod1.getElementsByTagName("retrigPhase");
       if (rpNodes.getLength() > 0) {
         try {
-          synth.setMod1RetrigPhase(Integer.parseInt(rpNodes.item(0).getTextContent().trim()));
+          synth.setMod1RetrigPhase((int) Long.parseLong(rpNodes.item(0).getTextContent().trim()));
         } catch (NumberFormatException e) {
           LOG.log(Level.FINE, "NumberFormatException parsing XML attribute", e);
         }
@@ -1845,7 +1845,7 @@ public class DelugeXmlParser {
       NodeList rpNodes = mod2.getElementsByTagName("retrigPhase");
       if (rpNodes.getLength() > 0) {
         try {
-          synth.setMod2RetrigPhase(Integer.parseInt(rpNodes.item(0).getTextContent().trim()));
+          synth.setMod2RetrigPhase((int) Long.parseLong(rpNodes.item(0).getTextContent().trim()));
         } catch (NumberFormatException e) {
           LOG.log(Level.FINE, "NumberFormatException parsing XML attribute", e);
         }
@@ -2358,7 +2358,7 @@ public class DelugeXmlParser {
       NodeList rpNodes = osc1El.getElementsByTagName("retrigPhase");
       if (rpNodes.getLength() > 0) {
         try {
-          sound.setOsc1RetrigPhase(Integer.parseInt(rpNodes.item(0).getTextContent().trim()));
+          sound.setOsc1RetrigPhase((int) Long.parseLong(rpNodes.item(0).getTextContent().trim()));
         } catch (NumberFormatException e) {
           LOG.log(Level.FINE, "NumberFormatException parsing XML attribute", e);
         }
@@ -2416,7 +2416,7 @@ public class DelugeXmlParser {
       NodeList rpNodes2 = osc2.getElementsByTagName("retrigPhase");
       if (rpNodes2.getLength() > 0) {
         try {
-          sound.setOsc2RetrigPhase(Integer.parseInt(rpNodes2.item(0).getTextContent().trim()));
+          sound.setOsc2RetrigPhase((int) Long.parseLong(rpNodes2.item(0).getTextContent().trim()));
         } catch (NumberFormatException e) {
           LOG.log(Level.FINE, "NumberFormatException parsing XML attribute", e);
         }
